Test name checking ...
