+++ src/SDL_internal.h
@@ -21,189 +21,464 @@
 #ifndef SDL_internal_h_
 #define SDL_internal_h_
 
-/* Many of SDL's features require _GNU_SOURCE on various platforms */
-#ifndef _GNU_SOURCE
-#define _GNU_SOURCE
-#endif
+#include "SDL_internal_common.h"
 
-/* Need this so Linux systems define fseek64o, ftell64o and off64_t */
-#ifndef _LARGEFILE64_SOURCE
-#define _LARGEFILE64_SOURCE 1
-#endif
+/* internal and internal lite are mutually exclusive */
+#define SDL_internal_lite_h_
 
-/* This is for a variable-length array at the end of a struct:
-    struct x { int y; char z[SDL_VARIABLE_LENGTH_ARRAY]; };
-   Use this because GCC 2 needs different magic than other compilers. */
-#if (defined(__GNUC__) && (__GNUC__ <= 2)) || defined(__CC_ARM) || defined(__cplusplus)
-#define SDL_VARIABLE_LENGTH_ARRAY 1
+
+#ifdef __HAIKU__
+/* nothing */
+#elif defined(__clang_analyzer__) && !defined(SDL_DISABLE_ANALYZE_MACROS)
+/* nothing */
 #else
-#define SDL_VARIABLE_LENGTH_ARRAY
-#endif
 
-#define SDL_MAX_SMALL_ALLOC_STACKSIZE          128
-#define SDL_small_alloc(type, count, pisstack) ((*(pisstack) = ((sizeof(type) * (count)) < SDL_MAX_SMALL_ALLOC_STACKSIZE)), (*(pisstack) ? SDL_stack_alloc(type, count) : (type *)SDL_malloc(sizeof(type) * (count))))
-#define SDL_small_free(ptr, isstack) \
-    if ((isstack)) {                 \
-        SDL_stack_free(ptr);         \
-    } else {                         \
-        SDL_free(ptr);               \
-    }
+#ifdef abs
+#undef abs
+#endif
+#define abs Please_use_SDL_abs_instead
+#ifdef acos
+#undef acos
+#endif
+#define acos Please_use_SDL_acos_instead
+#ifdef acosf
+#undef acosf
+#endif
+#define acosf Please_use_SDL_acosf_instead
+#ifdef asin
+#undef asin
+#endif
+#define asin Please_use_SDL_asin_instead
+#ifdef asinf
+#undef asinf
+#endif
+#define asinf Please_use_SDL_asinf_instead
+#ifdef asprintf
+#undef asprintf
+#endif
+#define asprintf Please_use_SDL_asprintf_instead
+#ifdef atan
+#undef atan
+#endif
+#define atan Please_use_SDL_atan_instead
+#ifdef atan2
+#undef atan2
+#endif
+#define atan2 Please_use_SDL_atan2_instead
+#ifdef atan2f
+#undef atan2f
+#endif
+#define atan2f Please_use_SDL_atan2f_instead
+#ifdef atanf
+#undef atanf
+#endif
+#define atanf Please_use_SDL_atanf_instead
+#ifdef atof
+#undef atof
+#endif
+#define atof Please_use_SDL_atof_instead
+#ifdef atoi
+#undef atoi
+#endif
+#define atoi Please_use_SDL_atoi_instead
+#ifdef calloc
+#undef calloc
+#endif
+#define calloc Please_use_SDL_calloc_instead
+#ifdef ceil
+#undef ceil
+#endif
+#define ceil Please_use_SDL_ceil_instead
+#ifdef ceilf
+#undef ceilf
+#endif
+#define ceilf Please_use_SDL_ceilf_instead
+#ifdef copysign
+#undef copysign
+#endif
+#define copysign Please_use_SDL_copysign_instead
+#ifdef copysignf
+#undef copysignf
+#endif
+#define copysignf Please_use_SDL_copysignf_instead
+#ifdef cos
+#undef cos
+#endif
+#define cos Please_use_SDL_cos_instead
+#ifdef cosf
+#undef cosf
+#endif
+#define cosf Please_use_SDL_cosf_instead
+#ifdef crc32
+#undef crc32
+#endif
+#define crc32 Please_use_SDL_crc32_instead
+#ifdef exp
+#undef exp
+#endif
+#define exp Please_use_SDL_exp_instead
+#ifdef expf
+#undef expf
+#endif
+#define expf Please_use_SDL_expf_instead
+#ifdef fabs
+#undef fabs
+#endif
+#define fabs Please_use_SDL_fabs_instead
+#ifdef fabsf
+#undef fabsf
+#endif
+#define fabsf Please_use_SDL_fabsf_instead
+#ifdef floor
+#undef floor
+#endif
+#define floor Please_use_SDL_floor_instead
+#ifdef floorf
+#undef floorf
+#endif
+#define floorf Please_use_SDL_floorf_instead
+#ifdef fmod
+#undef fmod
+#endif
+#define fmod Please_use_SDL_fmod_instead
+#ifdef fmodf
+#undef fmodf
+#endif
+#define fmodf Please_use_SDL_fmodf_instead
+#ifdef getenv
+#undef getenv
+#endif
+#define getenv Please_use_SDL_getenv_instead
+#ifdef isalnum
+#undef isalnum
+#endif
+#define isalnum Please_use_SDL_isalnum_instead
+#ifdef isalpha
+#undef isalpha
+#endif
+#define isalpha Please_use_SDL_isalpha_instead
+#ifdef isblank
+#undef isblank
+#endif
+#define isblank Please_use_SDL_isblank_instead
+#ifdef iscntrl
+#undef iscntrl
+#endif
+#define iscntrl Please_use_SDL_iscntrl_instead
+#ifdef isdigit
+#undef isdigit
+#endif
+#define isdigit Please_use_SDL_isdigit_instead
+#ifdef isgraph
+#undef isgraph
+#endif
+#define isgraph Please_use_SDL_isgraph_instead
+#ifdef islower
+#undef islower
+#endif
+#define islower Please_use_SDL_islower_instead
+#ifdef isprint
+#undef isprint
+#endif
+#define isprint Please_use_SDL_isprint_instead
+#ifdef ispunct
+#undef ispunct
+#endif
+#define ispunct Please_use_SDL_ispunct_instead
+#ifdef isspace
+#undef isspace
+#endif
+#define isspace Please_use_SDL_isspace_instead
+#ifdef isupper
+#undef isupper
+#endif
+#define isupper Please_use_SDL_isupper_instead
+#ifdef isxdigit
+#undef isxdigit
+#endif
+#define isxdigit Please_use_SDL_isxdigit_instead
+#ifdef itoa
+#undef itoa
+#endif
+#define itoa Please_use_SDL_itoa_instead
+#ifdef lltoa
+#undef lltoa
+#endif
+#define lltoa Please_use_SDL_lltoa_instead
+#ifdef log10
+#undef log10
+#endif
+#define log10 Please_use_SDL_log10_instead
+#ifdef log10f
+#undef log10f
+#endif
+#define log10f Please_use_SDL_log10f_instead
+#ifdef logf
+#undef logf
+#endif
+#define logf Please_use_SDL_logf_instead
+#ifdef lround
+#undef lround
+#endif
+#define lround Please_use_SDL_lround_instead
+#ifdef lroundf
+#undef lroundf
+#endif
+#define lroundf Please_use_SDL_lroundf_instead
+#ifdef ltoa
+#undef ltoa
+#endif
+#define ltoa Please_use_SDL_ltoa_instead
+#ifdef malloc
+#undef malloc
+#endif
+#define malloc Please_use_SDL_malloc_instead
+#ifdef memcmp
+#undef memcmp
+#endif
+#define memcmp Please_use_SDL_memcmp_instead
+#ifdef memcpy
+#undef memcpy
+#endif
+#define memcpy Please_use_SDL_memcpy_instead
+#ifdef memcpy4
+#undef memcpy4
+#endif
+#define memcpy4 Please_use_SDL_memcpy4_instead
+#ifdef memmove
+#undef memmove
+#endif
+#define memmove Please_use_SDL_memmove_instead
+/* no memset */
+#ifdef pow
+#undef pow
+#endif
+#define pow Please_use_SDL_pow_instead
+#ifdef powf
+#undef powf
+#endif
+#define powf Please_use_SDL_powf_instead
+#ifdef qsort
+#undef qsort
+#endif
+#define qsort Please_use_SDL_qsort_instead
+#ifdef realloc
+#undef realloc
+#endif
+#define realloc Please_use_SDL_realloc_instead
+#ifdef round
+#undef round
+#endif
+#define round Please_use_SDL_round_instead
+#ifdef roundf
+#undef roundf
+#endif
+#define roundf Please_use_SDL_roundf_instead
+#ifdef scalbn
+#undef scalbn
+#endif
+#define scalbn Please_use_SDL_scalbn_instead
+#ifdef scalbnf
+#undef scalbnf
+#endif
+#define scalbnf Please_use_SDL_scalbnf_instead
+#ifdef setenv
+#undef setenv
+#endif
+#define setenv Please_use_SDL_setenv_instead
+#ifdef sin
+#undef sin
+#endif
+#define sin Please_use_SDL_sin_instead
+#ifdef sinf
+#undef sinf
+#endif
+#define sinf Please_use_SDL_sinf_instead
+#ifdef snprintf
+#undef snprintf
+#endif
+#define snprintf Please_use_SDL_snprintf_instead
+#ifdef sqrt
+#undef sqrt
+#endif
+#define sqrt Please_use_SDL_sqrt_instead
+#ifdef sqrtf
+#undef sqrtf
+#endif
+#define sqrtf Please_use_SDL_sqrtf_instead
+#ifdef sscanf
+#undef sscanf
+#endif
+#define sscanf Please_use_SDL_sscanf_instead
+#ifdef strcasecmp
+#undef strcasecmp
+#endif
+#define strcasecmp Please_use_SDL_strcasecmp_instead
+#ifdef strchr
+#undef strchr
+#endif
+#define strchr Please_use_SDL_strchr_instead
+#ifdef strcmp
+#undef strcmp
+#endif
+#define strcmp Please_use_SDL_strcmp_instead
+#ifdef strdup
+#undef strdup
+#endif
+#define strdup Please_use_SDL_strdup_instead
+#ifdef strlcat
+#undef strlcat
+#endif
+#define strlcat Please_use_SDL_strlcat_instead
+#ifdef strlcpy
+#undef strlcpy
+#endif
+#define strlcpy Please_use_SDL_strlcpy_instead
+#ifdef strlen
+#undef strlen
+#endif
+#define strlen Please_use_SDL_strlen_instead
+#ifdef strlwr
+#undef strlwr
+#endif
+#define strlwr Please_use_SDL_strlwr_instead
+#ifdef strncasecmp
+#undef strncasecmp
+#endif
+#define strncasecmp Please_use_SDL_strncasecmp_instead
+#ifdef strncmp
+#undef strncmp
+#endif
+#define strncmp Please_use_SDL_strncmp_instead
+#ifdef strrchr
+#undef strrchr
+#endif
+#define strrchr Please_use_SDL_strrchr_instead
+#ifdef strrev
+#undef strrev
+#endif
+#define strrev Please_use_SDL_strrev_instead
+#ifdef strstr
+#undef strstr
+#endif
+#define strstr Please_use_SDL_strstr_instead
+#ifdef strtod
+#undef strtod
+#endif
+#define strtod Please_use_SDL_strtod_instead
+#ifdef strtokr
+#undef strtokr
+#endif
+#define strtokr Please_use_SDL_strtokr_instead
+#ifdef strtol
+#undef strtol
+#endif
+#define strtol Please_use_SDL_strtol_instead
+#ifdef strtoll
+#undef strtoll
+#endif
+#define strtoll Please_use_SDL_strtoll_instead
+#ifdef strtoul
+#undef strtoul
+#endif
+#define strtoul Please_use_SDL_strtoul_instead
+#ifdef strupr
+#undef strupr
+#endif
+#define strupr Please_use_SDL_strupr_instead
+#ifdef tan
+#undef tan
+#endif
+#define tan Please_use_SDL_tan_instead
+#ifdef tanf
+#undef tanf
+#endif
+#define tanf Please_use_SDL_tanf_instead
+#ifdef tolower
+#undef tolower
+#endif
+#define tolower Please_use_SDL_tolower_instead
+#ifdef toupper
+#undef toupper
+#endif
+#define toupper Please_use_SDL_toupper_instead
+#ifdef trunc
+#undef trunc
+#endif
+#define trunc Please_use_SDL_trunc_instead
+#ifdef truncf
+#undef truncf
+#endif
+#define truncf Please_use_SDL_truncf_instead
+#ifdef uitoa
+#undef uitoa
+#endif
+#define uitoa Please_use_SDL_uitoa_instead
+#ifdef ulltoa
+#undef ulltoa
+#endif
+#define ulltoa Please_use_SDL_ulltoa_instead
+#ifdef ultoa
+#undef ultoa
+#endif
+#define ultoa Please_use_SDL_ultoa_instead
+#ifdef utf8strlcpy
+#undef utf8strlcpy
+#endif
+#define utf8strlcpy Please_use_SDL_utf8strlcpy_instead
+#ifdef utf8strlen
+#undef utf8strlen
+#endif
+#define utf8strlen Please_use_SDL_utf8strlen_instead
+#ifdef vasprintf
+#undef vasprintf
+#endif
+#define vasprintf Please_use_SDL_vasprintf_instead
+#ifdef vsnprintf
+#undef vsnprintf
+#endif
+#define vsnprintf Please_use_SDL_vsnprintf_instead
+#ifdef vsscanf
+#undef vsscanf
+#endif
+#define vsscanf Please_use_SDL_vsscanf_instead
+#ifdef wcscasecmp
+#undef wcscasecmp
+#endif
+#define wcscasecmp Please_use_SDL_wcscasecmp_instead
+#ifdef wcscmp
+#undef wcscmp
+#endif
+#define wcscmp Please_use_SDL_wcscmp_instead
+#ifdef wcsdup
+#undef wcsdup
+#endif
+#define wcsdup Please_use_SDL_wcsdup_instead
+#ifdef wcslcat
+#undef wcslcat
+#endif
+#define wcslcat Please_use_SDL_wcslcat_instead
+#ifdef wcslcpy
+#undef wcslcpy
+#endif
+#define wcslcpy Please_use_SDL_wcslcpy_instead
+#ifdef wcslen
+#undef wcslen
+#endif
+#define wcslen Please_use_SDL_wcslen_instead
+#ifdef wcsncasecmp
+#undef wcsncasecmp
+#endif
+#define wcsncasecmp Please_use_SDL_wcsncasecmp_instead
+#ifdef wcsncmp
+#undef wcsncmp
+#endif
+#define wcsncmp Please_use_SDL_wcsncmp_instead
+#ifdef wcsstr
+#undef wcsstr
+#endif
+#define wcsstr Please_use_SDL_wcsstr_instead
 
-#include "build_config/SDL_build_config.h"
-
-#include "dynapi/SDL_dynapi.h"
-
-#if SDL_DYNAMIC_API
-#include "dynapi/SDL_dynapi_overrides.h"
-/* force DECLSPEC off...it's all internal symbols now.
-   These will have actual #defines during SDL_dynapi.c only */
-#ifdef DECLSPEC
-#undef DECLSPEC
-#endif
-#define DECLSPEC
-#endif
-
-#ifdef __APPLE__
-#ifndef _DARWIN_C_SOURCE
-#define _DARWIN_C_SOURCE 1 /* for memset_pattern4() */
-#endif
-#endif
-
-#ifdef HAVE_SYS_TYPES_H
-#include <sys/types.h>
-#endif
-#ifdef HAVE_STDIO_H
-#include <stdio.h>
-#endif
-#ifdef HAVE_STDLIB_H
-#include <stdlib.h>
-#elif defined(HAVE_MALLOC_H)
-#include <malloc.h>
-#endif
-#ifdef HAVE_STDDEF_H
-#include <stddef.h>
-#endif
-#ifdef HAVE_STDARG_H
-#include <stdarg.h>
-#endif
-#ifdef HAVE_STRING_H
-#ifdef HAVE_MEMORY_H
-#include <memory.h>
-#endif
-#include <string.h>
-#endif
-#ifdef HAVE_STRINGS_H
-#include <strings.h>
-#endif
-#ifdef HAVE_WCHAR_H
-#include <wchar.h>
-#endif
-#ifdef HAVE_INTTYPES_H
-#include <inttypes.h>
-#elif defined(HAVE_STDINT_H)
-#include <stdint.h>
-#endif
-#ifdef HAVE_CTYPE_H
-#include <ctype.h>
-#endif
-#ifdef HAVE_MATH_H
-#include <math.h>
-#endif
-#ifdef HAVE_FLOAT_H
-#include <float.h>
-#endif
-
-/* If you run into a warning that O_CLOEXEC is redefined, update the SDL configuration header for your platform to add HAVE_O_CLOEXEC */
-#ifndef HAVE_O_CLOEXEC
-#define O_CLOEXEC 0
-#endif
-
-/* A few #defines to reduce SDL footprint.
-   Only effective when library is statically linked.
-   You have to manually edit this file. */
-#ifndef SDL_LEAN_AND_MEAN
-#define SDL_LEAN_AND_MEAN 0
-#endif
-
-/* Optimized functions from 'SDL_blit_0.c'
-   - blit with source BitsPerPixel < 8, palette */
-#ifndef SDL_HAVE_BLIT_0
-#define SDL_HAVE_BLIT_0 !SDL_LEAN_AND_MEAN
-#endif
-
-/* Optimized functions from 'SDL_blit_1.c'
-   - blit with source BytesPerPixel == 1, palette */
-#ifndef SDL_HAVE_BLIT_1
-#define SDL_HAVE_BLIT_1 !SDL_LEAN_AND_MEAN
-#endif
-
-/* Optimized functions from 'SDL_blit_A.c'
-   - blit with 'SDL_BLENDMODE_BLEND' blending mode */
-#ifndef SDL_HAVE_BLIT_A
-#define SDL_HAVE_BLIT_A !SDL_LEAN_AND_MEAN
-#endif
-
-/* Optimized functions from 'SDL_blit_N.c'
-   - blit with COLORKEY mode, or nothing */
-#ifndef SDL_HAVE_BLIT_N
-#define SDL_HAVE_BLIT_N !SDL_LEAN_AND_MEAN
-#endif
-
-/* Optimized functions from 'SDL_blit_N.c'
-   - RGB565 conversion with Lookup tables */
-#ifndef SDL_HAVE_BLIT_N_RGB565
-#define SDL_HAVE_BLIT_N_RGB565 !SDL_LEAN_AND_MEAN
-#endif
-
-/* Optimized functions from 'SDL_blit_AUTO.c'
-   - blit with modulate color, modulate alpha, any blending mode
-   - scaling or not */
-#ifndef SDL_HAVE_BLIT_AUTO
-#define SDL_HAVE_BLIT_AUTO !SDL_LEAN_AND_MEAN
-#endif
-
-/* Run-Length-Encoding
-   - SDL_SetSurfaceColorKey() called with SDL_RLEACCEL flag */
-#ifndef SDL_HAVE_RLE
-#define SDL_HAVE_RLE !SDL_LEAN_AND_MEAN
-#endif
-
-/* Software SDL_Renderer
-   - creation of software renderer
-   - *not* general blitting functions
-   - {blend,draw}{fillrect,line,point} internal functions */
-#ifndef SDL_VIDEO_RENDER_SW
-#define SDL_VIDEO_RENDER_SW !SDL_LEAN_AND_MEAN
-#endif
-
-/* YUV formats
-   - handling of YUV surfaces
-   - blitting and conversion functions */
-#ifndef SDL_HAVE_YUV
-#define SDL_HAVE_YUV !SDL_LEAN_AND_MEAN
-#endif
-
-#include <SDL3/SDL.h>
-#include <SDL3/SDL_intrin.h>
-
-#define SDL_MAIN_NOIMPL /* don't drag in header-only implementation of SDL_main */
-#include <SDL3/SDL_main.h>
-
-/* The internal implementations of these functions have up to nanosecond precision.
-   We can expose these functions as part of the API if we want to later.
-*/
-/* Set up for C function definitions, even when using C++ */
-#ifdef __cplusplus
-extern "C" {
-#endif
-
-extern DECLSPEC int SDLCALL SDL_WaitSemaphoreTimeoutNS(SDL_Semaphore *sem, Sint64 timeoutNS);
-extern DECLSPEC int SDLCALL SDL_WaitConditionTimeoutNS(SDL_Condition *cond, SDL_Mutex *mutex, Sint64 timeoutNS);
-extern DECLSPEC int SDLCALL SDL_WaitEventTimeoutNS(SDL_Event *event, Sint64 timeoutNS);
-
-/* Ends C function definitions when using C++ */
-#ifdef __cplusplus
-}
 #endif
 
 #endif /* SDL_internal_h_ */
