+++ src/audio/pipewire/SDL_pipewire.c
@@ -19,7 +19,7 @@
   3. This notice may not be removed or altered from any source distribution.
 */
 
-#include "SDL_internal.h"
+#include "SDL_internal_lite.h"
 
 #ifdef SDL_AUDIO_DRIVER_PIPEWIRE
