+++ src/hidapi/SDL_hidapi.c
@@ -27,7 +27,7 @@
  * This merges the two, at a small performance cost, until distributions
  * have granted access to /dev/hidraw*
  */
-#include "SDL_internal.h"
+#include "SDL_internal_lite.h"
 
 #include "SDL_hidapi_c.h"
 #include "../joystick/usb_ids.h"
