+++ src/hidapi/android/hid.cpp
@@ -18,7 +18,7 @@
      misrepresented as being the original software.
   3. This notice may not be removed or altered from any source distribution.
 */
-#include "SDL_internal.h"
+#include "SDL_internal_lite.h"
 
 // Purpose: A wrapper implementing "HID" API for Android
 //
