+++ src/libm/e_log.c
@@ -1,4 +1,4 @@
-#include "SDL_internal.h"
+#include "SDL_internal_lite.h"
 /*
  * ====================================================
  * Copyright (C) 1993 by Sun Microsystems, Inc. All rights reserved.
