+++ src/video/uikit/SDL_uikitappdelegate.m
@@ -18,7 +18,9 @@
      misrepresented as being the original software.
   3. This notice may not be removed or altered from any source distribution.
 */
-#include "SDL_internal.h"
+#include "SDL_internal_lite.h"
+/* use of malloc/strelen before SDL starts */
+
 
 #ifdef SDL_VIDEO_DRIVER_UIKIT
